Move service/plugin architecture classes to Framework namespace.
