+++ src/Orion/Items/OrionItemService.cs
@@ -21,6 +21,7 @@ using System.Diagnostics;
 using Orion.Collections;
 using Orion.Events;
 using Orion.Events.Items;
+using Orion.Framework;
 using Orion.Packets.DataStructures;
 using Serilog;
