+++ src/Orion/Players/OrionPlayerService.cs
@@ -29,6 +29,7 @@ using Orion.Events;
 using Orion.Events.Packets;
 using Orion.Events.Players;
 using Orion.Events.World.Tiles;
+using Orion.Framework;
 using Orion.Packets;
 using Orion.Packets.Client;
 using Orion.Packets.Modules;
