+++ src/Orion/World/OrionWorldService.cs
@@ -17,6 +17,7 @@
 
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using Orion.Framework;
 using Orion.World.Tiles;
 using Serilog;
