+++ tests/Orion.Tests/Framework/OrionPluginTests.cs
@@ -16,40 +16,38 @@
 // along with Orion.  If not, see <https://www.gnu.org/licenses/>.
 
 using System;
-using Moq;
 using Serilog;
 using Serilog.Core;
 using Xunit;
 
-namespace Orion {
-    public class OrionServiceTests {
+namespace Orion.Framework {
+    public class OrionPluginTests {
         [Fact]
         public void Ctor_NullKernel_ThrowsArgumentNullException() {
-            Assert.Throws<ArgumentNullException>(() => new TestService(null!, Logger.None));
+            Assert.Throws<ArgumentNullException>(() => new TestOrionPlugin(null!, Logger.None));
         }
 
         [Fact]
         public void Ctor_NullLog_ThrowsArgumentNullException() {
             using var kernel = new OrionKernel(Logger.None);
 
-            Assert.Throws<ArgumentNullException>(() => new TestService(kernel, null!));
+            Assert.Throws<ArgumentNullException>(() => new TestOrionPlugin(kernel, null!));
         }
 
         [Fact]
-        public void Log_Get() {
+        public void Kernel_Get() {
             using var kernel = new OrionKernel(Logger.None);
-            var mockLog = new Mock<ILogger>();
-            using var service = new TestService(kernel, mockLog.Object);
+            using var plugin = new TestOrionPlugin(kernel, Logger.None);
 
-            service.Log.Information("test");
-
-            mockLog.Verify(l => l.Information("test"));
+            Assert.Same(kernel, plugin.Kernel);
         }
 
-        public class TestService : OrionService {
-            public new ILogger Log => base.Log;
+        private class TestOrionPlugin : OrionPlugin {
+            public new OrionKernel Kernel => base.Kernel;
 
-            public TestService(OrionKernel kernel, ILogger log) : base(kernel, log) { }
+            public TestOrionPlugin(OrionKernel kernel, ILogger log) : base(kernel, log) { }
+
+            public override void Initialize() => throw new NotImplementedException();
         }
     }
 }
