+++ tests/Orion.Tests/Framework/ServiceAttributeTests.cs
@@ -19,7 +19,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using Xunit;
 
-namespace Orion {
+namespace Orion.Framework {
     [SuppressMessage("Style", "IDE0017:Simplify object initialization", Justification = "Testing")]
     public class ServiceAttributeTests {
         [Fact]
